move resolveRoot to configs/utils
